Correct parsing issues with regex during somatic_exome GCP runs
somatic_exome workflow was breaking with an error indicating that the
sample name wasn't being extracted from the bam properly. Replicated
in a local docker, and there are three issues that were happening
with these regexes.

1. Perl seems to be reading the samtools results as a single string,
not a collection of strings. The leading ^ to pin to start of string
no longer works, because they're several lines down.

2. Somehow filter_vcf_dom.cwl was missing the .+ to span the line

3. filter_vcf_dom needs to escape the @. Unsure why mutect doesn't?

Error:
> Possible unintended interpolation of @RG in string at /app/docm/docm_filter.pl line 16.
> Global symbol "@RG" requires explicit package name (did you forget to declare "my @RG"?) at /app/docm/docm_filter.pl line 16.